refactor(eval): split paired eval tests by domain
Replace the single 800-line test_paired_eval.py (kept under the U-16
ceiling by deleting a blank line in bbc4b1a) with a real decomposition:
removal/dataset/judge/real-execution modules plus a shared helper.
test_paired_eval.py stays as the aggregate entry point so the documented
'python3 eval/test_paired_eval.py' verify command still runs all 23
tests.
