+++ eval/paired_test_support.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+"""Shared helpers for the paired evaluation test modules."""
+
+from __future__ import annotations
+
+from unittest.mock import Mock
+
+
+def sequence_client(replies: list[str]) -> Mock:
+    client = Mock()
+    client.messages.create.side_effect = [
+        type("Response", (), {"content": [type("Block", (), {"text": reply})()]})()
+        for reply in replies
+    ]
+    return client
